requirements.txt
python-dotenv==1.0.1
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
pypdf==4.3.1
python-dotenv==1.0.1
typing_extensions==4.12.2
typing-inspect==0.9.0
faiss-cpu==1.8.0
chromadb
streamlit
langchain_chroma
unstructured
#magic 오류 -> pip install git+https://github.com/julian-r/python-magic.git
#pip install -r requirements.txt